SW-26954 - Update phpstan/phpstan to 1.9.1

See merge request shopware/5/product/shopware!942

+++ _sql/migrations/795-remove-salutation-snippets.php
@@ -1,4 +1,6 @@
 <?php
+
+declare(strict_types=1);
 /**
  * Shopware 5
  * Copyright (c) shopware AG
@@ -22,13 +24,19 @@
  * our trademarks remain entirely with us.
  */
 
-class Migrations_Migration795 extends Shopware\Components\Migrations\AbstractMigration
+use Shopware\Components\Migrations\AbstractMigration;
+
+class Migrations_Migration795 extends AbstractMigration
 {
     public function up($modus)
     {
         $statement = $this->connection->query('SELECT id, locale_id FROM s_core_shops WHERE `default` = 1 LIMIT 1');
         $shop = $statement->fetchAll(PDO::FETCH_ASSOC);
         $shop = array_shift($shop);
+        if (!\is_array($shop)) {
+            return;
+        }
+
         $sql = "DELETE FROM s_core_snippets WHERE dirty = 0 AND namespace = 'frontend/salutation' AND value = '' AND (shopID != " . (int) $shop['id'] . ' OR localeID != ' . (int) $shop['locale_id'] . ')';
         $this->addSql($sql);
     }
